Have `edit_list` return an enum which indicates when the left and right values are completely unrelated.
Previously, the fallback option when the edit distance between the two arguments to `edit_list` exceeded the maximum was to return `ExtraLeft` and `ExtraRight` values and ignore any common values. This would be a problem when using the `contains`, `starts_with`, and `ends_with` matchers, since the edit list could easily exceed the maximum on large inputs.

This change replaces the output of `edit_list` with an enum which explicitly indicates whether a list of edits could be found. If it cannot, then the output generator in `eq_matcher` suppresses the diff output entirely.

PiperOrigin-RevId: 539567235