Ensure that the diff output of `starts_with` contains the last line of the actual value corresponding to the expected value when they are not equal.
Previously, the displayed diff from `starts_with` was suboptimal when the last line of the prefix differs from its corresponding line in the actual value (which occurs often):

Actual:

```
 First line
 Second line
 Third line
```

Prefix:

```
 First line
 Second
```

Diff:
```
 First line
-Second
<---- remaining lines omitted ---->
```

The line

```
+Second line
```

is missing from the diff.

This adds a hack which manually introduces the missing line to the diff in the above case. Doing this manually should not result in any incorrect output -- at worst, it could result in an additional line of output which was not strictly necessary.

PiperOrigin-RevId: 540592869