Better support the creation of a diff between actual and expected values with the `ends_with` matcher.
This uses the new `Prefix` mode in `edit_list` to skip the lines preceding the expected suffix. To do this, `StrMatcher` reverses the lines of the actual and expected values before creating the edit list, then restores the correct order from the output.

This approach has the side effect that the order in which the actual and expected lines are shown in the diff is the reverse of that used in the `eq` and `starts_with` matcher. It is unlikely that this difference will have any real consequence for the user, so I am leaving it as is.

PiperOrigin-RevId: 540859950