Improve support for displaying the diff between actual and expected values in the `contains_substring` matcher.
In contrast with `starts_with` and `ends_with`, there is no anchor to identify where to start matching the substring with the actual value. Any method to find such an anchor in general would require using a maximum edit distance equal to the sum of the lengths of the inputs. The resulting algorithm would have quadratic running time in the input length, which is too expensive.

For this reason, this implementation makes a "best effort" attempt using the existing maximum edit length and exact matching. The `edit_list` implementation has a new mode `Contains` which compresses the prefix and suffix in the actual value to `Edit::AdditionalLeft`. If the total number of lines in the actual value not in the substring already exceed the maximum edit distance, then the diff output will not be shown.

The practical effect of this change is that the lines of the actual value preceding and succeeding the substring are now compressed to the string `<---- remaining lines omitted ---->`.

PiperOrigin-RevId: 540904195