Disable all abbreviation of displayed values if the environment variable `DISABLE_ABBREVIATION` is set to any value.
This provides a stopgap if the abbreviation of displayed values is preventing the diagnosis of a test failure. This can be the case if:

 * The actual and expected values are strings containing newlines,
 * The actual and expected values differ by so much that the diff is not shown, and
 * The difference is not clear from the first and last 30 characters.

This combination of factors is not expected to occur often, but the user now has the chance to set an environment variable to see the full actual and expected values.

PiperOrigin-RevId: 542845889