Add LICENSE symlinks
Without this, `cargo publish` will upload a tarball that has no
LICENSE file. This in turn makes the get_rust_pkg.py script in Android
error out when trying to import the crate to AOSP.

The symlinks is all that is needed: Cargo will automatically pack up
the content (not the symlink) when publishing.