Introduce a matcher `char_count` to match string slices and owned strings with a particular character count.
Previously, there was no matcher on just the length of a string. The existing matcher `len` only works on containers, not strings.

The concept of "length of a string" is itself slightly ambiguous -- it could mean byte length, character count, or number of grapheme clusters. This matcher unambiguously counts characters, which should be correct in most cases.