Replace the `tuple!` macro with an implementation of `Matcher` for tuples.
This means that one can just put a bunch of matchers into a tuple and the result will itself be a matcher against corresponding tuples:

```
let value = (1, 2);
verify_that!(value, (eq(1), eq(2)))
```

There is no need for a separate macro for this, since one can implement `Matcher` for a tuple of matchers directly.

This is a breaking change, since it eliminates the `tuple!` macro. To port existing code, one just removes the call to `tuple!` and adds a trailing comma if necessary.