Clarify the behaviour of fatal and non-fatal assertions in the documentation.
This adjusts the text to make it clear that using the ? operator on a `verify_that!` result only causes the test to abort if *that assertion* fails and not if there are previous failing non-fatal assertions. It also adds a further example to illustrate the relationship between fatal and non-fatal assertions.