Use the supports-color crate to determine whether to produce coloured diffs.
This outsources some of the tricky logic for determining whether to produce ANSI sequences into a separate crate which serves as a source of truth.

This also removes the ANSI colour sequences from nearly all tests. It creates two integration tests which specifically test the colourised output respectively the behaviour of the environment variable `NO_COLOR`. Otherwise, it globally sets `NO_COLOR` to suppress colourised output. This greatly reduces the noise in the existing test assertions and makes sure the tests continue to be compatible with all environments.

The environment variable `GTEST_RUST_NO_COLOR` is hereby no longer supported. Instead the more standard variable `NO_COLOR` has the same effect. See <http://no-color.org>.

Due to one of the transitive dependencies of the supports-color crate, this raises the minimum supported Rust version of this crate to 1.64.