Remove the reexports of the individual matcher functions from the `matchers` module.
This moves the list of reexports of the individual matchers to the `prelude` module. Doing so reduces the number of public paths to each matcher function from three to two (the function itself and its reexport in the prelude), which makes the API surface of the library simpler.

This is a breaking change, since matchers can no longer be imported from the `matchers` module but must be imported from their specific submodules.

Downstream users using the prelude to access matchers will continue to be able to do so.