Support test functions returning () with `#[googletest::test]`.
A test which only uses the `expect_*` family of assertions (thus requiring `#[googletest::test]`) will only ever return `Ok(())`. There is therefore no reason to require the function itself to return a `Result`. That requirement just creates unnecessary boilerplate.

This change modifies the `#[googletest::test]` implementation to remove the requirement that the return type be present. If it is present, then it must still be a `Result`.

The generated compiler error can no longer occur and is removed. It would be ideal to check whether the given return type is in fact a `Result` and return a compiler error if it is not, but that is beyond the scope of this change.