Rewrite README and crate-level documentation to give a frontline summary of the `*_that!` assertion macros.
This should make the existence of and difference between these macros clearer up front to the reader. The `assert_that!` and `expect_that!` macros are listed first to nudge users to those rather than `verify_that!`.

This change also adds more examples of assertions in the beginning of the top-level documentation and changes the examples in crate-level documentation to use `expect_that!` rather than `verify_that!` when it does not otherwise matter.

Also fixes the Tokio dependency in the integration tests to 1.29.1, since the just released 1.32.0 appears to have a higher MSRV than this crate.

PiperOrigin-RevId: 558057297