Fix the language of the documentation of the type alias `Result`.
The existing language stated that test functions *must* return `Result<()>`. This is not true in light of the `assert_*` family of macros and is no longer true when using `expect_*` in combination with `#[googletest::test]`. This change updates the docstring to indicate that a test function *may* return the type alias for fatal assertion failures and that some of the assertion macros evaluate to it.