Use the `displays_as` matcher when asserting on the output of `Matcher::describe` in tests.
Many existing tests assert directly on the output of `Matcher::describe`, which is currently a `String`. This return type will change to a `Description` in the future. In order that the tests do not have to be modified when that is changed, this changes them to wrap the existing matcher in `displays_as`. That has no effect on the current assertion, but will cause the `Display::fmt` method to be invoked on the output in the future.