Change `Matcher::describe` and `Matcher::explain_match` to return `Description` rather than `String`.
Previously, the description and match explanation were rendered greedily as a `String`. This meant that each matcher was responsible for making its own local decisions about how its content was rendered, such as whether and where newlines should be placed. But each matcher does not have enough context to make such decisions, leading to various corner cases where rendering was suboptimal.

This will allow the rendering of the description to be controlled more precisely once all of its components are known.

This change only changes the return values of the above methods and of certain helper methods. It should be effectively a no-op, since it just packs the existing rendered `String`'s into `Description` structs.

This also introduces a couple of trait implementations for `Description` to simplify construction of a `Description` from strings as well as from sequences of other `Description`.

Future commits will change how the `Description` are constructed to take advantage of the structure now being communicated upwards.