Rewrite `Description` to use the new description renderer introduced in the previous commit.
Unit tests of `Description` are adjusted to use more idiomatic APIs.

This also changes the pre-existing direct users of `Description` to use idiomatic APIs rather than producing a `String` and wrapping that in a `Description` as they did before. Otherwise, this change would introduce several regressions in these matcher's explanations.

This removes the now unused method `Description::indent_except_first_line`.

Finally, the changes introduced earlier to indent the diff output for the `eq` and related matchers are now removed, since the indentation is now handled by the renderer.