Introduce a new toolset to render `Description`.
This new renderer is intended to replace the current `Display` implementation of `Description`. Currently it is not hooked up. This commit only introduces the rendering code and tests of it.

In the follow-up, `Description` will be reworked so that it holds structs from this new renderer. It will construct those structs on the fly as the `Description` is built. This new architecture decouples the construction of `Description` by the individual matchers from the rendering process.