Indent the output of the diff between actual and expected values by two spaces.
This is to reduce the need to adjust tests in a later change. Namely, a coming rework in the way failure messages are rendered will increase the aforementioned indentation, which will result in several tests having to be updated. Making this change now greatly reduces the number of tests which have to be updated later, allowing that change to be more focused.