Introduce a matcher for byte sequences which represent UTF-8 encoded strings.
This allows the use of all matchers for string types on such types. These types come up often when working with various types of I/O, so matching on them can be useful.

This currently only supports UTF-8 encoding, but could be extended for other encodings with the addition of a suitable (presumably optional) dependency.