Update the documentation of `Description` to reflect recent changes.
The previous documentation was written when `Description` was just a helper. This new documentation is geared to its new, more central role.

Also adjusts the trait bounds on some trait implementations for `Description` and `List` to allow iterators of string slices to be collected into a `Description`.

Also fixes two examples in the `Matcher` documentation.