Fix the tests on Rust nightly.
A recent change in Rust caused minor formatting changes when displaying fully qualified enum variants. Namely:
* The spaces around the `::` which were previously present were removed.
* The space between `vec!` and the vector content was removed.
This breaks tests which assert descriptions, match explanations, and failure messages. These rely on the `Debug` representation of actual and expected values.

This fixes the tests so that they pass again in nightly while continuing to pass in the other versions. It removes text from the assertion where it is not crucial to what is being tested. In cases where it is crucial, either the critical part is extracted to a variable which is set according to the Rust version, or the test itself is duplicated and adjusted according to the Rust version.