Rename `googletest::test` to `googletest::gtest`, and make `test` an alias for `gtest`
As we were adopting Googletest Rust within Google, we received feedback that
`#[googletest::test]` is too long, inconvenient to type, that it repeats the
word "test" twice, or that it is too much detail for a user to care about.

Thus, we decided to rename `#[googletest::test]` to `#[gtest]`, also adding it
to the prelude. This is consistent with existing practice in other Rust crates
that enhance the testing experience, for example, `#[rstest]`.

This commit starts a multi-commit chain.

We are not deprecating `#[googletest::test]`, because this spelling is useful
for compatibility with the rstest crate.

PiperOrigin-RevId: 660991306