Make `expect_eq` and friends more hygienic
Currently, `expect_that!` doesn't work unless both `verify_eq!` and `verify_that` are in scope.

This is okay or users of the prelude, but not for those who don't.

I've noticed that googletest-rust has a number of other macro hygiene issue, but to avoid making an overly large PR, I'll address them as I come across them.