Make `matches_pattern` for braced structs support exhaustiveness with generic structs.
Previously, for field exhaustiveness, we emitted an isolated `Foo { a: _, b: _ }` divorced from the `actual` value, but if `Foo`'s `a` field is generic, we're not able to infer it at all. So instead, put the `match` within a `predicate` matcher and match against the `actual` value to be able to infer the field.

Toward #447

PiperOrigin-RevId: 705913111