Relax trait bound restrictions for more flexible use of match_matrix.
This relaxes the earlier type constraint of requiring the `actual` collection to be `Copy` or `Debug`.  The main reason this was done was to use the `count_elements` method, but this can be avoided by building the graph with nested `map` uses instead.

The main benefit is to allow the `MatchMatrix` to accept a wider class of inputs, which can be helpful in the definition of new matchers.  A simliar relaxation is done to `get_explanation`, which also doesn't depend on it.

We also add a unit test to exercise the basic construction of the graph and expose some private functions to the crate in anticipation of its use in new matchers.

PiperOrigin-RevId: 707632959