Rename `into_test_result` to `or_fail`
This name is shorter and aligns with the existing `or_` methods
on `Option` and `Result`. It reads more clearly when used in
situations like `let error = my_result.err().or_fail()?;`

This CL also removes IntoTestResult's generic parameter.
All types that implement IntoTestResult do so for only one particular
output type, so the generic parameter is unnecessary.

NOTE: this is a backwards-incompatible change.
PiperOrigin-RevId: 712700561