Make the gtest macro compatible with no_std crates.
The code emitted by the gtest macro now references Result from
the core crate rather than the std crate. Avoiding the reference
to the std crate makes the resulting code no_std compatible.

The Result struct was moved from std to core in Rust v1.6.0, which
is part of the Rust 2018 edition.