Support negated patterns with `-` and change from `:` to `,` for pattern separation.
We adjust the supported pattern syntax to include negated patterns, improving alignment with [GTest C++](https://google.github.io/googletest/advanced.html#running-a-subset-of-the-tests).

Unfortunately, we have to deviate from GTest C++ with respect to the pattern separator `:` for ergonomic reasons.  The colon character `:` doesn't work well to separate the patterns because `::` is Rust's namespace separator.  This clash makes it much too easy to make unintentional mistakes when trying to write filter patterns with module paths.

To avoid this conflict, we switch to `,` as the pattern separator.

PiperOrigin-RevId: 734611957