Fix the parsing logic of `matches_pattern!` to support `_` at the top level in top-level structs.
In the current parsing logic, the `_` field in pattern `Some(_)` is parsed as `None`, and as a result no FieldMatcher is generated, and the internal `is()` matcher only evaluates the innerMatcher(the field_matcher).

We add `pattern_only()` matcher for the case when field_matcher is empty and skip the `is()` wrapping.

Toward #626

PiperOrigin-RevId: 755284260