Fix the parsing logic of matches_pattern! to support `_` at the top level in braced enums.
In the current parsing logic, the `_` field in pattern `AsEnum::Foo { a_field : _ }` is parsed as None, and as a result no FieldMatcher is generated, and the internal `is()` matcher only evaluates the innerMatcher(the field_matcher).

With the current change, when no field_matcher is present, the matches pattern macro verifies the variant of the Enum with `pattern_only()` matcher.

Toward #626

PiperOrigin-RevId: 755381147