Use UFCS for `with_message` assertions
Before this CL, the assertion macro variants which accept messages
would add the message via a call to the `with_message` method from the
`GoogleTestSupport` trait. However, that trait may not be imported if
users are not bringing in the full googletest prelude.

This CL switches those method calls to UFCS and adds a utility method
`and_log_failure_with_message` in order to make such calls more readable.

PiperOrigin-RevId: 775593149