Bug fix: when Fixtures fail, tests should fail too.
For a test which accepts a Fixture and returns (), if the Fixture::set_up
returned an error then the test would not be evaluated and automatically
succeed. This is dangerous because the unnoticed failure of a fixture is
silent. The critical code is this:

    let result: #invocation_result_type = #invocation;
    TestOutcome::close_current_test_outcome(#result)

...where #result not the "result" binding from one line up but rather is
defined earlier as quote! {googletest::Result::Ok(())} when the test
function returns ReturnType::Default. Thus is "result", the actual
result of calling #invocation, thrown away. The problem does not happen
if the test function returns googletest::Result (or any other error type
that can be coerced into that).

Taking note that closure_body() already unconditionally wraps all test
bodies that return () in another block that returns googletest::Result
(before potentially wrapping further in Fixtures), the solution is to
just always use quote! {result}.

Also make the documentation more explicit about it: in the case of a
Fixture returning an error it only said that the test is "not evaluated",
not whether it succeeds or fails. But I think it can't have been
intentional that it succeeds.

PiperOrigin-RevId: 788557182